// dashboard/src/utils/hasura-api/generated/schemas/dropActionOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DropActionArgs } from './dropActionArgs';
import type { DropActionOperationType } from './dropActionOperationType';

export interface DropActionOperation {
  type: DropActionOperationType;
  args: DropActionArgs;
}