// dashboard/src/utils/hasura-api/generated/schemas/migrationRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { MigrationStep } from './migrationStep';

export interface MigrationRequest {
  /** Name of the migration */
  name: string;
  /** Steps to execute for the migration. Always provide a down migration as well when creating a migration that modifies the schema */
  up: MigrationStep[];
  /** Steps to execute to revert the migration. Always provide this when creating a migration that modifies the schema */
  down?: MigrationStep[];
  /** The data source to use */
  datasource: string;
  /** Whether to skip the execution of the migration */
  skip_execution?: boolean;
}